Organize into Pages / Components
