+++ src/GoalList.tsx
@@ -1,4 +1,4 @@
-import './GoalList.css';
+import "./GoalList.css";
 import { FiCheck } from "react-icons/fi";
 
 type GoalListProps = {
@@ -9,32 +9,33 @@ type GoalListProps = {
 };
 
 function GoalList(props: GoalListProps) {
-  const {closeClicked, goalSelected, goalUnselected, goalSlugs} = props;
+  const { closeClicked, goalSelected, goalUnselected, goalSlugs } = props;
 
   return (
     <div className="GoalList">
-      <div className="close-button"
-        onClick={() => closeClicked()}
-      >
+      <div className="close-button" onClick={() => closeClicked()}>
         <span>close</span>
         <FiCheck></FiCheck>
       </div>
-        {
-        goalSlugs?.map(name => 
+      {goalSlugs?.map(name => (
         <div key={name}>
-          <input type="checkbox" id={name} name={name} onChange={(value) => {
-            if (value.target.checked) {
-              goalSelected(name)
-            } else {
-              goalUnselected(name)
-            }
-          }}>
-          </input>
+          <input
+            type="checkbox"
+            id={name}
+            name={name}
+            onChange={value => {
+              if (value.target.checked) {
+                goalSelected(name);
+              } else {
+                goalUnselected(name);
+              }
+            }}
+          ></input>
           <label htmlFor={name}>{name}</label>
-        </div>)
-      }
+        </div>
+      ))}
     </div>
-  )
+  );
 }
 
 export default GoalList;
